Update code 1
This function takes three arguments: an operation (a string that can be "add", "subtract", "multiply", or "divide"), and two numbers. It uses an if-else statement to determine which operation to perform based on the value of the operation argument, and then returns the result. You can use the console.log() function to test the function by passing it example arguments.